Added More Parameters
N: Specifies the number of completions to generate for each prompt. The default is 1.

Stop: A string that terminates the completion.

PresencePenalty: A float value that penalizes model for producing token sequences that are too similar to the original prompt. The default is 0.

BestOf: An integer value that selects the N best completions from among several completions generated for each prompt. The default is 1.

Temperature: A float value that controls the randomness of the completion. A value of 0 will produce only the most likely next token, while a value of 1 will produce a more diverse set of completions. The default is 1.

TopP: A float value between 0 and 1 that specifies the cumulative probability mass that each completion should have. The default is 1, meaning that all tokens are included in the completions.

RepetitionPenalty: A float value that penalizes model for producing repeated tokens, making it less likely to repeat itself. The default is 1.

FrequencyPenalty: A float value that penalizes the model for producing tokens that are very frequent in the training data, making it less likely to generate common words. The default is 0.

SelectionMethod: A string value that specifies the method for selecting the next token in the completion. The options are "greedy" and "nucleus", with "greedy" being the default.